agent/requirements.txt
google-generativeai
fastapi
uvicorn
requests
python-dotenv
